+++ TM4C129E_DriverLib/xDriver_MCU/ADC/Driver/Comparator/Control/xSource/ADC_Comparator_TriggerCondition.c
@@ -43,7 +43,7 @@ ADC_nERROR ADC_Comparator__enSetTriggerConditionsByMask(ADC_nMODULE enModuleArg,
         enErrorReg = ADC_enERROR_OK;
         while(0U != u32ComparatorMaskReg)
         {
-            if(0UL != (ADC_enCOMPMASK_0 & u32ComparatorMaskReg))
+            if(0UL != ((uint32_t) ADC_enCOMPMASK_0 & u32ComparatorMaskReg))
             {
                 enErrorReg = ADC_Comparator__enSetTriggerConditionsByNumber(enModuleArg, (ADC_nCOMPARATOR) u32ComparatorReg, enTriggerConditionsArg);
             }
