+++ TM4C129E_DriverLib/xDriver_MCU/ADC/Driver/Sample/xSource/ADC_Sample_CompSelection.c
@@ -43,7 +43,7 @@ ADC_nERROR ADC_Sample__enSetComparatorByMask(ADC_nMODULE enModuleArg, ADC_nSEQMA
         enErrorReg = ADC_enERROR_OK;
         while(0U != u32SequencerMaskReg)
         {
-            if(0UL != (ADC_enSEQMASK_0 & u32SequencerMaskReg))
+            if(0UL != ((uint32_t) ADC_enSEQMASK_0 & u32SequencerMaskReg))
             {
                 enErrorReg = ADC_Sample__enSetComparatorByNumber(enModuleArg, (ADC_nSEQUENCER) u32SequencerReg, enSampleArg, enComparatorArg);
             }
